# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Monkey Planner API
  version: 1.0.0
  description: |
    REST API for Monkey Planner — a kanban/list board with issues, comments,
    custom properties, webhooks, and a calendar stats view.

    ### Status transition rules
    - Issues start as **Pending**.
    - `POST /api/issues/{id}/approve` is the only way to move from Pending → Approved.
    - `PATCH /api/issues/{id}` with `status: "Approved"` returns **409** (`use_approve_endpoint`).
    - Once Approved, free movement between Approved ↔ InProgress ↔ Done is allowed via PATCH.
    - Returning to Pending from any state is forbidden.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/boards/{boardId}/issues/reorder:
    post:
      tags: [Issues]
      summary: Reorder issues on a board
      description: |
        Reassigns `position` values to the supplied issue IDs in the given order.
        All listed issues must belong to the board.
      operationId: reorderIssues
      parameters:
        - $ref: '#/components/parameters/PathBoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueIds]
              properties:
                issueIds:
                  type: array
                  items:
                    type: string
                  description: Ordered list of issue IDs
                  example: ["id-3", "id-1", "id-2"]
      responses:
        '204':
          description: Reorder applied (no content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/issues:
    get:
      tags: [Issues]
      summary: List issues
      description: |
        Returns issues filtered by optional query parameters.
        - `board_id`: filter by board
        - `status`: filter by status value (`Pending`, `Approved`, `InProgress`, `Done`)
        - `parent_id`: filter by parent issue ID; pass an empty string to list only root issues
      operationId: listIssues
      parameters:
        - name: board_id
          in: query
          schema:
            type: string
          description: Filter by board ID
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/Status'
          description: Filter by status
        - name: parent_id
          in: query
          schema:
            type: string
          description: >
            Filter by parent issue ID.
            Omit to return all issues regardless of hierarchy.
            Pass empty string (`parent_id=`) to return only root-level issues (no parent).
      responses:
        '200':
          description: Array of issues (empty array when none match)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Issues]
      summary: Create an issue
      description: |
        Creates a new issue with status `Pending`.
        If `parentId` is supplied the issue becomes a child of that issue.
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boardId, title]
              properties:
                boardId:
                  type: string
                  example: "board-uuid"
                parentId:
                  type: string
                  nullable: true
                  description: Parent issue ID (omit for root-level issue)
                  example: "parent-uuid"
                title:
                  type: string
                  example: "Implement login screen"
                body:
                  type: string
                  example: "Use OAuth2 with Google provider"
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/issues/{id}:
    get:
      tags: [Issues]
      summary: Get an issue with its children
      operationId: getIssue
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Issue and its direct child issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  issue:
                    $ref: '#/components/schemas/Issue'
                  children:
                    type: array
                    items:
                      $ref: '#/components/schemas/Issue'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Issues]
      summary: Update an issue
      description: |
        Partial update — only provided fields are changed.

        **Special field: `properties`**
        When the `properties` key is present the update is handled as a
        property-merge operation and the response is returned immediately;
        other fields in the same request body are ignored.

        **Special field: `parentId`**
        - Omit the key: parent is unchanged.
        - Set to `null`: move issue to root level.
        - Set to a string ID: reparent to that issue.

        **Status rules**
        Setting `status` to `"Approved"` returns **409**
        (`use_approve_endpoint`). Use `POST /api/issues/{id}/approve` instead.
        Moving to `Pending` from any state is also forbidden.
      operationId: patchIssue
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated title"
                body:
                  type: string
                  example: "Updated description"
                status:
                  $ref: '#/components/schemas/Status'
                parentId:
                  type: string
                  nullable: true
                  description: >
                    New parent ID, null to move to root, or omit to leave unchanged.
                  example: null
                properties:
                  type: object
                  additionalProperties: true
                  description: >
                    Custom property values keyed by property ID.
                    When present, all other fields in this request are ignored
                    and a merge update is applied.
                  example:
                    prop-uuid-priority: "High"
                    prop-uuid-due: "2025-12-31"
      responses:
        '200':
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Conflict.
            - `use_approve_endpoint`: tried to set status to Approved via PATCH.
            - `conflict`: storage-level conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Issues]
      summary: Delete an issue
      operationId: deleteIssue
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Issue deleted (no content)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/issues/{id}/approve:
    post:
      tags: [Issues]
      summary: Approve an issue
      description: |
        Transitions the issue from **Pending** to **Approved** and records
        `approvedAt`. This is the only permitted path for this transition.

        The operation is idempotent: calling it on an already-Approved issue
        preserves the original `approvedAt` timestamp.
      operationId: approveIssue
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Approved issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Conflict — issue is not in a state that can be approved
            (e.g. already Approved, InProgress, or Done and the transition
            is meaningless, or a storage conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Comments ──────────────────────────────────────────────────────────────
  /api/issues/{issueId}/comments:
    get:
      tags: [Comments]
      summary: List comments for an issue
      operationId: listComments
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
          description: Issue ID
      responses:
        '200':
          description: Array of comments (empty array when none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Comments]
      summary: Create a comment on an issue
      operationId: createComment
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
          description: Issue ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  example: "Looks good to me!"
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/comments/{commentId}:
    delete:
      tags: [Comments]
      summary: Delete a comment
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
          description: Comment ID
      responses:
        '204':
          description: Comment deleted (no content)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Calendar ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
    PathBoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
      description: Board ID
  responses:
    BadRequest:
      description: Bad request — invalid input or JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Status:
      type: string
      enum: [Pending, Approved, InProgress, Done]
      description: Issue lifecycle status
      example: Pending
...
    Issue:
      type: object
      required: [id, boardId, title, body, status, properties, position, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: "i1a2b3c4-d5e6-7890-abcd-ef1234567890"
        boardId:
          type: string
          example: "b1a2c3d4-e5f6-7890-abcd-ef1234567890"
        parentId:
          type: string
          nullable: true
          description: Parent issue ID; null for root-level issues
          example: null
        title:
          type: string
          example: "Implement login screen"
        body:
          type: string
          example: "Use OAuth2 with Google provider"
        status:
          $ref: '#/components/schemas/Status'
        properties:
          type: object
          additionalProperties: true
          description: Custom property values keyed by property ID
          example:
            prop-uuid-priority: "High"
            prop-uuid-due: "2025-12-31"
        position:
          type: integer
          description: Sort position within the board column
          example: 0
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-02T12:00:00Z"
        approvedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the issue was approved; null until approved
          example: null
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the issue moved to Done; null until completed
          example: null
    Comment:
      type: object
      required: [id, issueId, body, createdAt]
      properties:
        id:
          type: string
          example: "c1a2b3c4-d5e6-7890-abcd-ef1234567890"
        issueId:
          type: string
          example: "i1a2b3c4-d5e6-7890-abcd-ef1234567890"
        body:
          type: string
          example: "Looks good to me!"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-03T09:00:00Z"
...
    ErrorResponse:
      type: object
      required: [error]
      description: Standard error envelope returned on all non-2xx responses
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "not_found"
            message:
              type: string
              description: Human-readable error description
              example: "not found"